LangGraph/requirements.txt
gradio==5.29.1
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.60
langchain_experimental==0.3.4
langchain_openai==0.3.17
langgraph==0.4.5
playwright==1.52.0
pydantic==2.11.4
python-dotenv==1.1.0
Requests==2.32.3
typing_extensions==4.13.2

